Add files via upload
This Geodemographic segmentation model was made with the purpose of determining customers who are most likely to to leave the bank. using the logistic regression algorithm. the model was improved by making use of the backwards elimination method, followed by testing the performance of the model via Cumulative Accuracy Profile(CAP) curve and the confusion matrix along with the classification report.